delete flask app, move all training2 django files to training
