+++ training/training/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for training2 project.
+ASGI config for training project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "training2.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "training.settings")
 
 application = get_asgi_application()
