+++ training/training/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for training2 project.
+Django settings for training project.
 
 Generated by 'django-admin startproject' using Django 4.2.3.
 
@@ -49,7 +49,7 @@ MIDDLEWARE = [
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
 
-ROOT_URLCONF = "training2.urls"
+ROOT_URLCONF = "training.urls"
 
 TEMPLATES = [
     {
@@ -67,7 +67,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = "training2.wsgi.application"
+WSGI_APPLICATION = "training.wsgi.application"
 
 
 # Database
