+++ training/training/trainer.py
@@ -6,7 +6,7 @@ from sklearn.metrics import confusion_matrix, roc_auc_score, roc_curve
 import torch
 from torch.utils.data import DataLoader
 import torch.nn as nn
-from training2.criterion import CriterionHandler
+from training.criterion import CriterionHandler
 
 
 @dataclass
